Agregada funcionalidades para registrar viajes, solicitar asientos etc, corregidos fallos en los modelos vehiculo y usuario. (Faltaria definir reglas del negocio)

+++ src/main/java/co/com/cidenet/backendpae/repository/UserRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface UserRepository extends MongoRepository<User,String> {
     List<User> findAll();
     User findUserById(String id);
-    User findUserByDocument(int document);
+    User findUserByDocument(long document);
     List<User> findUsersByFirstname(String firstname);
     User findUserByEmail(String email);
